requirements.txt
alembic==1.13.2
APIFlask==2.2.0
apispec==6.6.1
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-HTTPAuth==4.8.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.3
packaging==24.1
PyJWT==2.9.0
python-dotenv==1.0.1
SQLAlchemy==2.0.32
typing_extensions==4.12.2
webargs==8.4.0
Werkzeug==3.0.3
